New folder/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
Flask-Mail
Flask-Migrate
Werkzeug
email-validator
reportlab
